fix: max hold for various corner cases
max hold needs to work, even when the data is always negative, but this
complicates things because the ypoint array is initialized to zeros
prior to the first push of data.  The new code will wait until an entire
frame of data is pushed (since you can push partial frames) and then
initialize max hold to -Infinity.